feat: move to advanced plugin and add reset API
