+++ package.json
@@ -21,7 +21,7 @@
   "scripts": {
     "prebuild": "npm run clean",
     "build": "npm-run-all -p build:*",
-    "build:css": "node-sass src/plugin.scss dist/videojs-overlay.css --output-style=compressed --linefeed=lf",
+    "build:css": "sass --load-path=node_modules src/plugin.scss dist/videojs-overlay.css --style compressed",
     "build:js": "rollup -c scripts/rollup.config.js",
     "clean": "shx rm -rf ./dist ./test/dist",
     "postclean": "shx mkdir -p ./dist ./test/dist",
@@ -51,13 +51,13 @@
     "husky": "^1.0.0-rc.13",
     "karma": "^3.0.0",
     "lint-staged": "^8.1.0",
-    "node-sass": "^6.0.1",
     "not-prerelease": "^1.0.1",
     "npm-merge-driver-install": "^1.0.0",
     "npm-run-all": "^4.1.5",
     "pkg-ok": "^2.2.0",
     "postcss-cli": "^6.0.0",
     "rollup": "^2.61.1",
+    "sass": "^1.79.5",
     "shx": "^0.3.2",
     "sinon": "^6.1.5",
     "videojs-generate-karma-config": "^8.0.1",
@@ -80,6 +80,12 @@
     "src/",
     "test/"
   ],
+  "exports": {
+    "./plugin-only": {
+      "require": "./dist/videojs-overlay.plugin.js",
+      "import": "./dist/videojs-overlay.plugin.js"
+    }
+  },
   "generator-videojs-plugin": {
     "version": "7.3.2"
   },
